requirements.txt
# AI Video Editing Pipeline Dependencies

# Core dependencies
pydantic>=2.0.0,<3.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0

# ElevenLabs SDK
elevenlabs>=1.0.0

# Video editing
moviepy>=1.0.3

# HTML parsing
beautifulsoup4>=4.12.0

# Audio analysis
librosa>=0.10.0
numpy>=1.24.0

# Optional: for better type hints and validation
typing-extensions>=4.8.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# mypy>=1.5.0

